refactor: move 4 providers to domain dirs
Move log_provider, script_provider, update_provider, introspection_provider
out of include/ros2_medkit_gateway/providers/ into their domain directories
(logs/, scripts/, updates/, discovery/). Updates 24 #include lines across
gateway sources, tests, plugins, and docs.

Leaves the plugin-facing per-entity provider interfaces (data_provider,
fault_provider, operation_provider) in place - those serve a different
layer (SOVD plugin delegation) and their reorganization is a separate
design decision deferred to a follow-up.

Refs #375
