+++ src/ros2_medkit_gateway/include/ros2_medkit_gateway/script_manager.hpp
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "ros2_medkit_gateway/providers/script_provider.hpp"
+#include "ros2_medkit_gateway/scripts/script_provider.hpp"
 
 namespace ros2_medkit_gateway {
