+++ src/ros2_medkit_gateway/test/test_log_manager.cpp
@@ -22,9 +22,9 @@
 #include <thread>
 
 #include "ros2_medkit_gateway/log_manager.hpp"
+#include "ros2_medkit_gateway/logs/log_provider.hpp"
 #include "ros2_medkit_gateway/plugins/gateway_plugin.hpp"
 #include "ros2_medkit_gateway/plugins/plugin_manager.hpp"
-#include "ros2_medkit_gateway/providers/log_provider.hpp"
 #include "ros2_medkit_gateway/resource_change_notifier.hpp"
 
 using json = nlohmann::json;
